move methods onto EntityBinder and CollectionBinder
